requirements.txt
requests==2.31.0
django==3.2.20
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
PyJWT==2.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter~=21.1
python-dotenv==0.20.0